Create a component for calendar & add dots for meetings
