+++ pages/dashboard.vue
@@ -1,24 +1,9 @@
 <script setup lang="ts">
 	import MeetingCard from '@/components/MeetingCard.vue'
-	import type { Meeting } from '~/types'
-	import type {} from '@samk-dev/nuxt-vcalendar'
 
 	const { getMeetings } = useApi()
 
 	const { state: podState } = usePodStore()
-
-	const date = ref(new Date())
-
-	const attrs = ref<any>([
-		{
-			key: 'today',
-			highlight: {
-				color: 'green',
-				fillMode: 'solid',
-			},
-			dates: new Date(),
-		},
-	])
 
 	const { data, refresh, pending } = await useAsyncData('dashboard', async () => {
 		const [meetings] = await Promise.all([
@@ -57,11 +42,7 @@
 				<div class="col-span-1 grid gap-4 self-start lg:col-span-3">
 					<div class="rounded-lg bg-zinc-900 p-4">
 						<h2 class="mb-4 text-lg font-medium text-white">Upcoming events</h2>
-						<ClientOnly>
-							<div class="my-calendar">
-								<VCalendar v-model="date" :attributes="attrs" class="" isDark />
-							</div>
-						</ClientOnly>
+						<MeetingCalendar :meetings="data?.meetings" />
 						<div class="mt-4">
 							<ScheduleMeetingCard @refresh="refresh" />
 						</div>
